Add initial crud/base view prototype

+++ transport/src/main/java/czort/response/UserResponse.java
@@ -1,15 +1,15 @@
 package czort.response;
 
 public class UserResponse {
-    private Integer id;
+    private Long id;
     private String name;
     private String email;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -32,7 +32,7 @@ public class UserResponse {
     public UserResponse() {
     }
 
-    public UserResponse(Integer id, String name, String email) {
+    public UserResponse(Long id, String name, String email) {
         this.id = id;
         this.name = name;
         this.email = email;
